requirements.txt
# --- Orchestration & Configuration ---
PyYAML==6.0.3
pydantic==2.12.5

# --- Google Cloud SDK ---
google-cloud-bigquery==3.40.0
google-cloud-storage==3.1.1

# --- dbt (Core + Adapter BigQuery) ---
dbt-core==1.11.4
dbt-bigquery==1.11.0
